kzg_rs

Module kzg_proof

Structs

KzgProof

Functions

batch_inversion

Montgomery batch inversion in a finite field Given a list of elements ( x_1, x_2, \dots, x_n ) from a finite field ( F ), Montgomery batch inversion computes the inverses ( x_1^{-1}, x_2^{-1}, \dots, x_n^{-1} ) as follows:

compute_challenge

Return the Fiat-Shamir challenge required to verify blob and commitment.

compute_challenges_and_evaluate_polynomial

compute_powers

compute_r_powers

evaluate_polynomial_in_evaluation_form

Evaluates a polynomial in evaluation form at a given point

safe_g1_affine_from_bytes

safe_scalar_affine_from_bytes

scalar_from_bytes_unchecked

scalar_from_u64_array_unchecked

validate_batched_input

verify_kzg_proof_impl